CM-84 Rename SetAccountPassword to CreateAccount
For better compatibility with ChronoMint Web
