+++ src/registerScreens.js
@@ -24,18 +24,18 @@ import AddTimeLockedWallet from './containers/AddTimeLockedWalletContainer'
 import AddTokenToAdvancedWallet from './containers/AddTokenToAdvancedWalletContainer'
 import AddWallet from './containers/AddWalletContainer'
 import ConfirmMnemonic from './containers/ConfirmMnemonicContainer'
+import CreateAccount from './containers/CreateAccountContainer'
 import Download2FAApp from './containers/Download2FAAppContainer'
 import Drawer from './containers/DrawerContainer'
-import LoginWithMnemonic from './containers/LoginWithMnemonicContainer'
 import EnterPin from './containers/EnterPinContainer'
 import EnterPrivateKey from './containers/EnterPrivateKeyContainer'
 import GenerateMnemonic from './containers/GenerateMnemonicContainer'
 import LoginScreenLayout from './components/LoginScreenLayout'
+import LoginWithMnemonic from './containers/LoginWithMnemonicContainer'
 import screenLayout from './utils/screenLayout'
 import SelectAccount from './containers/SelectAccountContainer'
 import SelectLanguage from './containers/SelectLanguageContainer'
 import SelectNetwork from './containers/SelectNetworkContainer'
-import SetAccountPassword from './containers/SetAccountPasswordContainer'
 import WalletBackup from './containers/WalletBackupContainer'
 //#endregion
 
@@ -65,16 +65,16 @@ export default function registerScreens (store: TStore<TState, { type: string }>
   Navigation.registerComponent('AddTokenToAdvancedWallet', () => AddTokenToAdvancedWallet, store, Provider)
   Navigation.registerComponent('AddWallet', () => AddWallet, store, Provider)
   Navigation.registerComponent('ConfirmMnemonic', () => screenLayout(LoginScreenLayout)(ConfirmMnemonic), store, Provider)
+  Navigation.registerComponent('CreateAccount', () => screenLayout(LoginScreenLayout)(CreateAccount), store, Provider)
   Navigation.registerComponent('Download2FAApp', () => Download2FAApp, store, Provider)
   Navigation.registerComponent('Drawer', () => Drawer, store, Provider)
-  Navigation.registerComponent('LoginWithMnemonic', () => screenLayout(LoginScreenLayout)(LoginWithMnemonic), store, Provider)
   Navigation.registerComponent('EnterPin', () => screenLayout(LoginScreenLayout)(EnterPin), store, Provider)
   Navigation.registerComponent('EnterPrivateKey', () => screenLayout(LoginScreenLayout)(EnterPrivateKey), store, Provider)
   Navigation.registerComponent('GenerateMnemonic', () => screenLayout(LoginScreenLayout)(GenerateMnemonic), store, Provider)
+  Navigation.registerComponent('LoginWithMnemonic', () => screenLayout(LoginScreenLayout)(LoginWithMnemonic), store, Provider)
   Navigation.registerComponent('SelectAccount', () => screenLayout(LoginScreenLayout)(SelectAccount), store, Provider)
   Navigation.registerComponent('SelectLanguage', () => SelectLanguage, store, Provider)
   Navigation.registerComponent('SelectNetwork', () => SelectNetwork, store, Provider)
-  Navigation.registerComponent('SetAccountPassword', () => screenLayout(LoginScreenLayout)(SetAccountPassword), store, Provider)
   //#endregion
 
   //#region Wallet screens
